+++ Assets/Scripts/Weapon/WeaponStats.cs
@@ -5,7 +5,7 @@ public class WeaponStats : ScriptableObject
 {
     public float LifeTime;
     public float Speed;
-    public float Damage;
+    public int Damage;
     public float MoveAccuracy;
     public float StopAccuracy;
 }
